LayoutTests/canvas/philip/tests/2d.fillStyle.parse.current.removed-expected.txt

< [index] >
2d.fillStyle.parse.current.removed
currentColor is solid black when the canvas element is not in a document
References: 2d.colours.parse 2d.currentColor.outofdoc
Actual output:
Expected output:

Passed
